+++ .vercel/output/_functions/pages/_image.astro.mjs
@@ -1,2 +1,2 @@
-export { a as page } from '../chunks/generic_Avb3YuZH.mjs';
+export { a as page } from '../chunks/generic_CqrVGNmH.mjs';
 export { renderers } from '../renderers.mjs';
