+++ .vercel/output/static/_astro/FooterSocialLinks.DMsWxV6d.js
@@ -1,0 +1,1 @@
+import{j as e}from"./jsx-runtime.D_zvdyIk.js";import{S as t,a as r,b as a,c as l,d as i}from"./index.CkhJJBex.js";import"./index.DYrVU9rO.js";function x(){return e.jsxs("div",{className:"flex items-center justify-center space-x-6 flex-wrap gap-4",children:[e.jsx("a",{href:"https://github.com/N95Ryan",target:"_blank",rel:"noopener noreferrer","aria-label":"GitHub",className:"text-white hover:text-purple-400 transition-colors duration-300",children:e.jsx(t,{className:"text-2xl md:text-3xl",title:"GitHub"})}),e.jsx("a",{href:"https://www.linkedin.com/in/ryan-pina-silasse/",target:"_blank",rel:"noopener noreferrer","aria-label":"LinkedIn",className:"text-white hover:text-purple-400 transition-colors duration-300",children:e.jsx(r,{className:"text-2xl md:text-3xl",title:"LinkedIn"})}),e.jsx("a",{href:"https://x.com/BigNayru",target:"_blank",rel:"noopener noreferrer","aria-label":"X (Twitter)",className:"text-white hover:text-purple-400 transition-colors duration-300",children:e.jsx(a,{className:"text-2xl md:text-3xl",title:"X (Twitter)"})}),e.jsx("a",{href:"https://bento.me/nayru",target:"_blank",rel:"noopener noreferrer","aria-label":"Bento",className:"text-white hover:text-purple-400 transition-colors duration-300",children:e.jsx(l,{className:"text-2xl md:text-3xl",title:"Bento"})}),e.jsx("a",{href:"https://www.malt.fr/profile/ryanpinasilasse",target:"_blank",rel:"noopener noreferrer","aria-label":"Malt",className:"text-white hover:text-purple-400 transition-colors duration-300",children:e.jsx(i,{className:"text-2xl md:text-3xl",title:"Malt"})})]})}export{x as default};
