+++ .vercel/output/static/_astro/SkillsList.DXjt2z5C.js
@@ -1,11 +1,11 @@
-import{j as o}from"./jsx-runtime.D_zvdyIk.js";import{G as i,e as r,f as c,g as s,h as l,i as m,j as d,k as h,l as p,m as x,n as u,o as f,p as S,q as g,r as n,s as y,t as v,u as j,v as k,w as b,x as w,y as F,z as E,A as C,B as G}from"./index.CkhJJBex.js";import"./index.DYrVU9rO.js";function N({Icon:a,color:t}){return o.jsx(a,{style:{color:t},className:"text-8xl"})}function A({title:a,Icon:t,color:e}){return o.jsxs("div",{className:`
-        bg-gradient-to-br from-gray-800 to-gray-900 
-        hover:from-gray-700 hover:to-gray-800
-        p-4 text-white rounded-xl 
-        flex flex-col items-center justify-center space-y-2 
-        w-full sm:w-64 h-48
-        shadow-xl hover:shadow-2xl
-        transition-all duration-500 ease-out
-        hover:scale-105
-        backdrop-blur-sm
+import{j as o}from"./jsx-runtime.D_zvdyIk.js";import{G as i,e as r,f as c,g as s,h as l,i as m,j as d,k as h,l as p,m as x,n as u,o as f,p as S,q as g,r as n,s as y,t as v,u as j,v as k,w as b,x as w,y as F,z as E,A as C,B as G}from"./index.CkhJJBex.js";import"./index.DYrVU9rO.js";function N({Icon:a,color:t}){return o.jsx(a,{style:{color:t},className:"text-8xl"})}function A({title:a,Icon:t,color:e}){return o.jsxs("div",{className:`\r
+        bg-gradient-to-br from-gray-800 to-gray-900 \r
+        hover:from-gray-700 hover:to-gray-800\r
+        p-4 text-white rounded-xl \r
+        flex flex-col items-center justify-center space-y-2 \r
+        w-full sm:w-64 h-48\r
+        shadow-xl hover:shadow-2xl\r
+        transition-all duration-500 ease-out\r
+        hover:scale-105\r
+        backdrop-blur-sm\r
       `,style:{boxShadow:"0 10px 30px rgba(0, 0, 0, 0.3), 0 0 20px rgba(99, 102, 241, 0.1)"},children:[o.jsx("div",{className:"text-white text-lg sm:text-xl md:text-2xl font-semibold bg-gradient-to-r from-white to-gray-300 bg-clip-text text-transparent",children:a}),o.jsx("div",{className:"text-white",children:o.jsx(N,{Icon:t,color:e})})]})}function D(a){return i({attr:{viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},child:[{tag:"path",attr:{d:"M4 13h5"},child:[]},{tag:"path",attr:{d:"M12 16v-8h3a2 2 0 0 1 2 2v1a2 2 0 0 1 -2 2h-3"},child:[]},{tag:"path",attr:{d:"M20 8v8"},child:[]},{tag:"path",attr:{d:"M9 16v-5.5a2.5 2.5 0 0 0 -5 0v5.5"},child:[]}]})(a)}const B={javascript:{icon:G,color:"#F0DB4F",name:"JavaScript"},typescript:{icon:C,color:"#007acc",name:"TypeScript"},golang:{icon:E,color:"#00acd7",name:"Go"},python:{icon:F,color:"#ffcf3b",name:"Python"},react:{icon:n,color:"#61DAFB",name:"React"},nextjs:{icon:w,color:"#fff",name:"Next.js"},astro:{icon:b,color:"#e93dc9",name:"Astro"},tanstack:{icon:k,color:"#FF4154",name:"TanStack Query"},tailwind:{icon:j,color:"#06B6D4",name:"Tailwind CSS"},nodedotjs:{icon:v,color:"#539e43",name:"Node.js"},gin:{icon:y,color:"#00acd7",name:"Gin"},echo:{icon:D,color:"#1E90FF",name:"Echo"},reactNative:{icon:n,color:"#61DAFB",name:"React Native"},expo:{icon:g,color:"#fff",name:"Expo"},postgresql:{icon:S,color:"#336791",name:"PostgreSQL"},supabase:{icon:f,color:"#3ECF8E",name:"Supabase"},cloudinary:{icon:u,color:"#3448C5",name:"Cloudinary"},vite:{icon:x,color:"#fff000",name:"Vite"},graphql:{icon:p,color:"#E10098",name:"GraphQL"},postman:{icon:h,color:"#FF6C37",name:"Postman"},docker:{icon:d,color:"#2496ED",name:"Docker"},githubActions:{icon:m,color:"#2088FF",name:"GitHub Actions"},vercel:{icon:l,color:"#fff",name:"Vercel"},vitest:{icon:s,color:"#6E9F18",name:"Vitest"},playwright:{icon:c,color:"#2EAD33",name:"Playwright"},lighthouse:{icon:r,color:"#F44B21",name:"Lighthouse"}};function T({skills:a}){return o.jsx(o.Fragment,{children:a.map(t=>{const e=B[t.key];return e?o.jsx(A,{title:e.name,Icon:e.icon,color:e.color},t.key):(console.warn(`Skill config not found for key: ${t.key}`),null)})})}export{T as default};
