+++ src/layouts/BaseLayout.astro
@@ -79,7 +79,7 @@ const schemaData = {
     <meta name="description" content={description} />
     <meta
       name="keywords"
-      content="Ryan PINA-SILASSE, Software Engineer Toronto, React Developer Canada, Go Developer Toronto, Full Stack Developer Ontario, TypeScript Developer, Next.js Developer, Node.js Developer, Python Developer, React Native Developer, Mobile Developer Canada, Frontend Developer Toronto, Backend Engineer Toronto, Web Developer Quebec, Astro Developer, PostgreSQL Developer, Docker Engineer, GraphQL Developer, Tailwind CSS Developer, Supabase Developer, Cloudinary Developer, Canada, Freelance Developer Toronto, Remote Developer Canada, Software Engineer Quebec, Drummondville Developer, Available Developer Toronto, Contract Developer Ontario, Web Development Services Canada, React Expert Toronto, Go Lang Developer, TypeScript Expert, Full Stack Engineer Canada, Mobile App Developer, API Developer Toronto, Cloud Developer Canada"
+      content="Ryan PINA-SILASSE, Software Engineer Toronto, React Developer Canada, Go Developer Toronto, Golang Developer, Full Stack Developer Ontario, TypeScript Developer, Next.js Developer, Node.js Developer, Express Developer, Python Developer, React Native Developer, Expo Developer, Mobile Developer Canada, Frontend Developer Toronto, Backend Engineer Toronto, Web Developer Quebec, Astro Developer, PostgreSQL Developer, Docker Engineer, GraphQL Developer, Tailwind CSS Developer, Supabase Developer, Cloudinary Developer, Echo Framework, Gin Framework, Vercel, Canada, Freelance Developer Toronto, Remote Developer Canada, Software Engineer Quebec, Drummondville Developer, Available Developer Toronto, Contract Developer Ontario, Web Development Services Canada, React Expert Toronto, Go Lang Developer, TypeScript Expert, Full Stack Engineer Canada, Mobile App Developer, API Developer Toronto, Cloud Developer Canada"
     />
     <meta name="author" content="Ryan PINA-SILASSE" />
     <meta
